优化文档

+++ public/cpp/actor/actor_logic/actor_robot/message/actor_robot_handle.cpp
@@ -193,7 +193,7 @@ namespace ngl
 		auto lpram = adata.get_data();
 
 		i32_serverid lserverid = static_cast<i32_serverid>(lpram->mserverid());
-		int16_t ltcount = static_cast<int16_t>(nnodeid::tcount(lserverid));
+		int16_t ltcount = nnodeid::tcount(lserverid);
 		int16_t ltid = static_cast<int16_t>(nnodeid::tid(lserverid));
 
 		net_works lpstructserver;
